Add link flags for specific hostnames

+++ linkmanager/cplinks.py
@@ -1,5 +1,6 @@
 # encoding: utf-8
 import os, shutil
+from linkmanager import DELETED, HOSTNAME
 from linkmanager import log, utils
 from linkmanager import rmlink
 
@@ -12,23 +13,28 @@ def get_options(parser):
 
 def create_symlink(syncpath, home, linkroot, dryrun=False):
     """ Create a symlink for the specified syncpath. """
-    homepath = syncpath.replace(linkroot, home)
-    # if syncpath is a link, resolve its destination
-    if os.path.islink(syncpath):
-        syncpath = os.readlink(syncpath)
+    # check syncpath is flagged for a specific hostname or deleted
+    syncpath, syncflag = utils.get_syncflag(syncpath)
+    if syncflag is not None and syncflag == DELETED:
+        return rmlink.remove_syncpath(syncpath, linkroot, dryrun)
+    if syncflag is not None and syncflag != HOSTNAME:
+        return None
     # check the homepath file or dir already exists and delete it
     # TODO: we should prompt before deleting anything here
+    # WARNING: Do not put homepath before syncpath in the below code!
+    homepath = syncpath.replace(linkroot, home)
+    syncpath = os.readlink(syncpath) if os.path.islink(syncpath) else syncpath
     if os.path.exists(homepath) or os.path.islink(homepath):
         if os.path.islink(homepath) and os.readlink(homepath) == syncpath:
             log.debug(f'Existing link: {homepath}')
             return
-        log.info(f'Deleting: {homepath}')
+        log.debug(f'Deleting: {homepath}')
         if (os.path.isfile(homepath) or os.path.islink(homepath)) and not dryrun:
             os.remove(homepath)
         elif os.path.isdir(homepath) and not dryrun:
             shutil.rmtree(homepath)
     # make sure home dirs exiist and create the new symlink
-    log.info(f'Syncing: {homepath}')
+    log.info(f'Syncing: {homepath} -> {syncpath}')
     if not dryrun:
         os.makedirs(os.path.dirname(homepath), exist_ok=True)
         os.symlink(syncpath, homepath)
@@ -37,7 +43,4 @@ def create_symlink(syncpath, home, linkroot, dryrun=False):
 def run_command(opts):
     """ Symlink synced files and dirs to home directory. """
     for ftype, syncpath in utils.iter_linkroot(opts.linkroot):
-        if utils.is_deleted(syncpath):
-            rmlink.remove_syncpath(syncpath, opts.linkroot, opts.dryrun)
-            continue
         create_symlink(syncpath, opts.home, opts.linkroot, opts.dryrun)
